requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pymongo==4.10.1
python-dotenv==1.1.1
bcrypt==4.2.0
email-validator==2.2.0
